% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfofunc.R
\name{diff_pfo}
\alias{diff_pfo}
\title{Difference between two portfolios}
\usage{
diff_pfo(from, to)
}
\arguments{
\item{from}{initial portfolio.}

\item{to}{target portfolio.}
}
\value{
A named list of differences indexed by ticker.
}
\description{
Difference between two portfolios: difference from `from` to `to`.
}
\examples{
p1 <- pfo_init('USD', 100000)
p2 <- pfo_trade(p1, 'MSCI', 10.5, 1756)
p3 <- pfo_trade(p2, 'VIX', 20, 14)
from <- pfo_grepl(p1, 'MSCI|VIX')
to <- pfo_grepl(p3, 'MSCI|VIX')
diff_pfo(from, to)

}
