% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfofunc.R
\name{diff_trades}
\alias{diff_trades}
\title{Compute trades between a start and a target pfo}
\usage{
diff_trades(from, to, dt, val_pri.e)
}
\arguments{
\item{from}{initial portfolio.}

\item{to}{target portfolio.}

\item{dt}{date.}

\item{val_pri.e}{price values environment.}
}
\value{
A list of trades.
}
\description{
Compute trades between a start and a target pfo.
}
\examples{
p1 <- pfo_init('USD', 100000)
p2 <- pfo_trade(p1, 'MSCI', 10.5, 1756)
p3 <- pfo_trade(p2, 'VIX', 20, 14)
from <- pfo_grepl(p1, 'MSCI|VIX')
to <- pfo_grepl(p3, 'MSCI|VIX')
dt <- as.Date('2019-05-22')
xts_msci <- xts::xts(1756, order.by = dt)
xts_vix <- xts::xts(14, order.by = dt)
val_pri.l <- list('MSCI' = xts_msci, 'VIX' = xts_vix)
val_pri.e <- list2env(val_pri.l)
diff_trades(from, to, dt, val_pri.e)

}
