% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfofunc.R
\name{instr_price_on}
\alias{instr_price_on}
\title{Get price of instrument on specific date}
\usage{
instr_price_on(xts_pri_env, tic, dt)
}
\arguments{
\item{xts_pri_env}{environment of xts timeseries
indexed by ticker.}

\item{tic}{instrument ticker.}

\item{dt}{date.}
}
\value{
A number.
}
\description{
Get price of instrument on specific date. Extract price from
environment of timeseries.
}
\examples{
dt <- as.Date('2019-10-02')
tsmsci <- xts::xts(x = c(2121.47), dt)
xtslst <- list('MSCI' = tsmsci)
xtsenv <- list2env(xtslst)
instr_price_on(xtsenv, 'MSCI', dt)

}
