% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfofunc.R
\name{pfo_currency}
\alias{pfo_currency}
\title{Portfolio currency}
\usage{
pfo_currency(pfo)
}
\arguments{
\item{pfo}{portfoio.}
}
\value{
A string.
}
\description{
Portfolio currency.
}
\examples{
p1 <- pfo_init('USD', 100000)
pfo_currency(p1)

}
