% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfofunc.R
\name{pfo_from_list}
\alias{pfo_from_list}
\title{Initialize portfolio from list}
\usage{
pfo_from_list(currency, lst)
}
\arguments{
\item{currency}{three-letter currency (string).}

\item{lst}{list of instrument quantities indexed by instrument tickers.}
}
\value{
A portfolio.
}
\description{
Initialize portfolio with given currency from a list.
}
\examples{
pfo_from_list('USD', list('USD' = 13937.10, 'MSCI' = 12.08, 'VIX' = 20))

}
