% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfofunc.R
\name{pfo_grepl}
\alias{pfo_grepl}
\title{Get subset portfolio with instrument grepl}
\usage{
pfo_grepl(pfo, pattern)
}
\arguments{
\item{pfo}{input portfolio.}

\item{pattern}{regular expression string.}
}
\value{
A portfolio.
}
\description{
Get subset portfolio with instrument grepl.
}
\examples{
p1 <- pfo_init('USD', 100000)
p2 <- pfo_trade(p1, 'MSCI', 10.5, 1756)
pfo_grepl(p1, '^MSCI$')
pfo_grepl(p2, '^MSCI$')
pfo_grepl(p2, '.')

}
