% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfofunc.R
\name{pfo_init}
\alias{pfo_init}
\title{Initialize portfolio}
\usage{
pfo_init(currency, init_cash)
}
\arguments{
\item{currency}{three-letter currency (string).}

\item{init_cash}{initial cash (number).}
}
\value{
A portfolio.
}
\description{
Initialize portfolio with given currency and a cash amount.
}
\examples{
pfo_init('USD', 100000)

}
