% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfofunc.R
\name{pfo_instr_prices_on}
\alias{pfo_instr_prices_on}
\title{Get prices of instruments in pfo}
\usage{
pfo_instr_prices_on(pfo, dt, xts_val_pri.e)
}
\arguments{
\item{pfo}{input portfolio.}

\item{dt}{date.}

\item{xts_val_pri.e}{environment of xts timeseries
indexed by ticker.}
}
\value{
A named numeric vector.
}
\description{
Get prices of instruments in pfo
}
\examples{
p1 <- pfo_init('USD', 100000)
p2 <- pfo_trade(p1, 'MSCI', 10.5, 1756)
dt <- as.Date('2019-10-02')
tsmsci <- xts::xts(x = c(2121.47), dt)
xtslst <- list('MSCI' = tsmsci)
xtsenv <- list2env(xtslst)
pfo_instr_prices_on(p1, dt, xtsenv)
pfo_instr_prices_on(p2, dt, xtsenv)

}
