% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfofunc.R
\name{pfo_instr_qty}
\alias{pfo_instr_qty}
\title{Quantity of an instrument in a portfolio}
\usage{
pfo_instr_qty(pfo, instr)
}
\arguments{
\item{pfo}{input portfolio.}

\item{instr}{instrument identifier (string).}
}
\value{
A number.
}
\description{
Quantity of an instrument in a portfolio.
}
\examples{
p1 <- pfo_init('USD', 100000)
p2 <- pfo_trade(p1, 'MSCI', 10.5, 1756)
pfo_instr_qty(p2, 'USD')
pfo_instr_qty(p2, 'MSCI')
pfo_instr_qty(p2, 'EUR')

}
