% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfofunc.R
\name{pfo_instr_str}
\alias{pfo_instr_str}
\title{Description of pfo instruments in string}
\usage{
pfo_instr_str(pfo)
}
\arguments{
\item{pfo}{input portfolio.}
}
\value{
A string.
}
\description{
Description of portfolio instruments in string.
}
\examples{
p1 <- pfo_init('USD', 100000)
p2 <- pfo_trade(p1, 'MSCI', 10.5, 1756)
pfo_instr_str(p1)
pfo_instr_str(p2)

}
