% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfofunc.R
\name{pfo_instruments}
\alias{pfo_instruments}
\title{Instruments in a portfolio}
\usage{
pfo_instruments(pfo)
}
\arguments{
\item{pfo}{input portfolio.}
}
\value{
A character vector.
}
\description{
Instruments in a portfolio.
}
\examples{
p1 <- pfo_init('USD', 100000)
p2 <- pfo_trade(p1, 'MSCI', 10.5, 1756)
pfo_instruments(p1)
pfo_instruments(p2)

}
