% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfofunc.R
\name{pfo_instruments_except_pfo_ccy}
\alias{pfo_instruments_except_pfo_ccy}
\title{Instruments in a portfolio except pfo ccy}
\usage{
pfo_instruments_except_pfo_ccy(pfo)
}
\arguments{
\item{pfo}{input portfolio.}
}
\value{
A character vector.
}
\description{
Instruments in a portfolio except portfolio currency.
}
\examples{
p1 <- pfo_init('USD', 100000)
p2 <- pfo_trade(p1, 'MSCI', 10.5, 1756)
pfo_instruments_except_pfo_ccy(p1)
pfo_instruments_except_pfo_ccy(p2)

}
