% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfofunc.R
\name{pfo_make_trades}
\alias{pfo_make_trades}
\title{Make trades in a portfolio}
\usage{
pfo_make_trades(pfo, trade_list)
}
\arguments{
\item{pfo}{input portfolio.}

\item{trade_list}{trade list.}
}
\value{
A portfolio.
}
\description{
Make trades in portfolio from a list of trades.
}
\examples{
p1 <- pfo_init('USD', 100000)
trade_list <- list(
  list(instrument = 'MSCI', trade_qty = 10, price = 1756),
  list(instrument = 'MSCI', trade_qty = 15, price = 1700)
)
pfo_make_trades(p1, trade_list)

}
