% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfofunc.R
\name{pfo_nav}
\alias{pfo_nav}
\title{Calculate nav of pfo given price list/env}
\usage{
pfo_nav(pfo, pl)
}
\arguments{
\item{pfo}{input portfolio.}

\item{pl}{price list or environment.}
}
\value{
A number.
}
\description{
Calculate nav of portfolio given price list or environment.
}
\examples{
p1 <- pfo_init('USD', 100000)
p2 <- pfo_trade(p1, 'MSCI', 10.5, 1756)
pl <- list('MSCI' = 1798)
pfo_nav(p2, pl)

}
