% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfofunc.R
\name{pfo_nav_on}
\alias{pfo_nav_on}
\title{Compute nav of pfo on specific date}
\usage{
pfo_nav_on(pfo, dt, xts_val_pri.e)
}
\arguments{
\item{pfo}{input portfolio.}

\item{dt}{date.}

\item{xts_val_pri.e}{environment of xts timeseries
indexed by ticker.}
}
\value{
A number.
}
\description{
Compute nav of portfolio on specific date.
}
\examples{
p1 <- pfo_init('USD', 100000)
p2 <- pfo_trade(p1, 'MSCI', 10.5, 1756)
dt <- as.Date('2019-10-02')
tsmsci <- xts::xts(x = c(2121.47), dt)
xtslst <- list('MSCI' = tsmsci)
xtsenv <- list2env(xtslst)
pfo_nav_on(p1, dt, xtsenv)
pfo_nav_on(p2, dt, xtsenv)

}
