% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfofunc.R
\name{pfo_target_trade}
\alias{pfo_target_trade}
\title{Make a trade in portfolio to a target qty}
\usage{
pfo_target_trade(pfo, instrument, target_qty, price)
}
\arguments{
\item{pfo}{input portfolio.}

\item{instrument}{financial instrument (string).}

\item{target_qty}{target quantity (number).}

\item{price}{price in portfolio currency (number).}
}
\value{
A portfolio.
}
\description{
Make a trade in portfolio, given an instrument,
its target final quantity and the price of the instrument
in portfolio currency.
}
\examples{
p1 <- pfo_init('USD', 100000)
p2 <- pfo_trade(p1, 'MSCI', 10.5, 1756)
p3 <- pfo_trade(p2, 'MSCI', 12.5, 1789)
p4 <- pfo_trade(p3, 'MSCI', 0, 1793)

}
