% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfofunc.R
\name{pfo_trade}
\alias{pfo_trade}
\title{Make a trade in portfolio}
\usage{
pfo_trade(pfo, instrument, trade_qty, price)
}
\arguments{
\item{pfo}{input portfolio.}

\item{instrument}{financial instrument (string).}

\item{trade_qty}{trade quantity (number).}

\item{price}{price in portfolio currency (number).}
}
\value{
A portfolio.
}
\description{
Make a trade in portfolio, given instrument, traded quantity
and price in portfolio currency.
}
\examples{
p1 <- pfo_init('USD', 100000)
pfo_trade(p1, 'MSCI', 10.5, 1756)

}
